backend/requirements.txt
# Vena RAG Bot Production - Dependencies
# Python 3.12+ (Python 3.13 has package compatibility issues)

# Core RAG Framework
numpy>=1.26.0,<2.0.0  # Compatible with langchain 0.1.20 and Python 3.12
langchain==0.1.20  # Pin to avoid breaking changes from 1.x
langchain-openai==0.1.7  # Pin to avoid breaking changes
langchain-community==0.0.38  # Pin to avoid breaking changes
anthropic>=0.75.0  # For Claude support (optional - only needed if using Anthropic)

# Vector Database (Production)
qdrant-client==1.9.0

# OpenAI
openai>=1.30.1

# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic>=2.10.0,<3.0.0  # Latest with Python 3.13 support
pydantic-settings>=2.6.0,<3.0.0  # Latest with Python 3.13 support

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.1

# Caching
redis>=5.0.1

# Document Processing
python-dotenv>=1.0.1
tiktoken>=0.12.0  # Latest available version (0.12.0 is max)
pypdf>=4.0.1
python-docx>=1.1.0  # For Word document parsing

# Utilities
httpx>=0.27.0  # Latest with Python 3.13 support
aiohttp>=3.9.0  # For async load testing

# AWS S3 Support (for knowledge base hosting)
boto3>=1.34.0

# Testing (optional, for development)
pytest==7.4.4
pytest-asyncio==0.23.3

# Jira Integration (for CSV import script)
jira==3.5.2

# Authentication (JWT)
PyJWT>=2.8.0
bcrypt>=4.0.1,<5.0.0

# Rate Limiting
slowapi>=0.1.9

# Monitoring & Error Tracking
sentry-sdk[fastapi]>=2.0.0
